pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "glonorce-youtube-mcp"
version = "2.0.0"
description = "A Model Context Protocol (MCP) server for YouTube operations"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "mcp>=1.6.0",
    "yt-ts-extract>=1.0.0",
    "yt-info-extract",
]

[project.scripts]
youtube_mcp = "youtube_mcp.server:main"

[tool.hatch.build.targets.wheel]
packages = ["src/youtube_mcp"]

[dependency-groups]
dev = [
    "hatch>=1.14.1",
    "mcp[cli]>=1.6.0",
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=6.2.1",
    "python-dotenv>=1.0.0",
    "twine>=6.1.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
